% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_comments}
\alias{cell_line_comments}
\title{Find comments from cell-lines}
\usage{
cell_line_comments(cell_line, category = NULL)
}
\arguments{
\item{cell_line}{A node or nodeset referring to a cell-line or multiple cell-lines.}

\item{category}{A character vector specifying which categories to output.}
}
\value{
The comment(s) associated with the cell-lines.
}
\description{
Expected values for \code{category} are
\itemize{
  \item \code{Anecdotal},
  \item \code{Biotechnology},
  \item \code{Breed/subspecies},
  \item \code{Caution},
  \item \code{Characteristics},
  \item \code{Derived from metastatic site},
  \item \code{Derived from sampling site},
  \item \code{Discontinued},
  \item \code{Doubling time},
  \item \code{From},
  \item \code{Group},
  \item \code{Knockout cell},
  \item \code{Microsatellite instability},
  \item \code{Miscellaneous},
  \item \code{Misspelling},
  \item \code{Monoclonal antibody target},
  \item \code{Omics},
  \item \code{Part of},
  \item \code{Population},
  \item \code{Problematic cell line},
  \item \code{Registration},
  \item \code{Selected for resistance to},
  \item \code{Sequence variation},
  \item \code{Transfected with},
  \item \code{Transformant}
}
}
\examples{
cellosaurus <- read_cellosaurus_xml("data/cellosaurus.xml")
mice_lines <- cell_line_find_all(cellosaurus, "Mus musculus")
cell_line_comments(mice_lines, category = "Anecdotal")

}
