% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_find_first}
\alias{cell_line_find_first}
\title{Find cell-line}
\usage{
cell_line_find_first(cellosaurus, text)
}
\arguments{
\item{cellosaurus}{An XML document containing the Cellosaurus dataset.}

\item{text}{Some text to search for. Finding is handled using XPath so some
special characters may cause difficulties, including these: '/:[]*.}
}
\description{
Find cell-line
}
\examples{
cellosaurus <- read_cellosaurus_xml("data/cellosaurus.xml")
cell_line_find_first(cellosaurus, "CVCL_3449")
# Can store the cell-line for use later.
CVCL_6873 <- cell_line_find_first(cellosaurus, "CVCL_6873")

}
