% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_names}
\alias{cell_line_names}
\title{Find names from a cell-line}
\usage{
cell_line_names(cell_line, type = NULL)
}
\arguments{
\item{cell_line}{A node or nodeset referring to a cell-line.}

\item{type}{A character vector specifying which types of names to output.
Expected values are \code{c("identifier", "synonym")}. If \code{NULL},
the function returns all names (i.e., unfiltered).}
}
\value{
The name(s) associated with the cell-line.
}
\description{
Find names from a cell-line
}
\examples{
cellosaurus <- read_cellosaurus_xml("data/cellosaurus.xml")
a_line <- cell_line_find_first(cellosaurus, "CVCL_E548")

# If type is not set, returns all names
cell_line_names(a_line)

# The expected values of type are "identifier" or "synonym":
cell_line_names(a_line, type = "identifier")
cell_line_names(a_line, type = "synonym")

# Can provide multiple type values as a character vector.
# This should be the same as the unfiltered version:
cell_line_names(a_line, type = c("identifier", "synonym"))

}
