% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_names_all}
\alias{cell_line_names_all}
\alias{cell_line_names_identifier}
\alias{cell_line_names_synonym}
\title{Find names from a cell-line}
\usage{
cell_line_names_all(cell_line)

cell_line_names_identifier(cell_line)

cell_line_names_synonym(cell_line)
}
\description{
\code{cell_line_names_all} finds the name entries from a
cell-line's name-list. Does not filter based on \code{type} so
may include entries marked as types both \code{identifier} and
\code{synonym}.

\code{cell_line_names_identifier} finds the name entries from a
cell-line's name-list, filtered to only include \code{type="identifier"}.
We expect that to only return one element, but do not enforce it.

\code{cell_line_names_synonym} finds the name entries from a cell-line's
name-list, filtered to only include \code{type="synonym"}.
}
\section{Functions}{
\itemize{
\item \code{cell_line_names_identifier}: Find identifier name(s) for a cell-line

\item \code{cell_line_names_synonym}: Find synonym names for a cell-line
}}

\examples{
cellosaurus <- read_cellosaurus_xml("data/cellosaurus.xml")
CVCL_E548 <- cell_line_find_first(cellosaurus, "CVCL_E548")
cell_line_names_all(CVCL_E548)
cell_line_names_synonym(CVCL_E548)
cell_line_names_identifier(CVCL_E548)

}
