% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_references}
\alias{cell_line_references}
\title{Find references from cell-lines}
\usage{
cell_line_references(cell_line)
}
\arguments{
\item{cell_line}{A node or nodeset referring to a cell-line or multiple cell-lines.}
}
\value{
The resource-internal-ref(s) associated with the cell-lines. These
  relate to the publication-list also stored in the Cellosaurus XML data.
}
\description{
Find references from cell-lines
}
\examples{
cellosaurus <- read_cellosaurus_xml("data/cellosaurus.xml")
a_line <- cell_line_find_first(cellosaurus, "CVCL_E548")
cell_line_webpages(a_line)

}
