% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_sex}
\alias{cell_line_sex}
\title{Sex of a cell-line}
\usage{
cell_line_sex(cell_line)
}
\description{
Sex of a cell-line
}
\examples{
# First read the dataset in.
cellosaurus <- read_cellosaurus_xml("data/cellosaurus.xml")
# Can call on a nodeset that's already been found. E.g.:
CVCL_E548 <- cell_line_find_first(cellosaurus, "CVCL_E548")
cell_line_sex(CVCL_E548)
# Or could nest the sex with the the finder:
cell_line_sex(cell_line_find_first(cellosaurus, "CVCL_3449"))
}
