% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_webpages}
\alias{cell_line_webpages}
\title{Find web pages from cell-lines}
\usage{
cell_line_webpages(cell_line)
}
\arguments{
\item{cell_line}{A node or nodeset referring to a cell-line or multiple cell-lines.}
}
\value{
The web page(s) associated with the cell-lines.
}
\description{
Find web pages from cell-lines
}
\examples{
cellosaurus <- read_cellosaurus_xml("data/cellosaurus.xml")
a_line <- cell_line_find_first(cellosaurus, "CVCL_E548")
cell_line_webpages(a_line)

}
