% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_lines_all}
\alias{cell_lines_all}
\title{Get all cell-line elements from Cellosaurus xml dataset}
\usage{
cell_lines_all(cellosaurus)
}
\arguments{
\item{cellosaurus}{An XML document containing the Cellosaurus dataset.}
}
\value{
An XML nodeset containing all cell-lines in the dataset.
}
\description{
Having read a Cellosaurus dataset with \code{\link{read_cellosaurus_xml}},
this function extracts all the cell-line elements as an \code{xml_nodeset}.
}
\details{
Once you have used this function, you will typically want to use
\code{\link{cell_lines_filter}} to find the cell-lines that are of interest
to you.
}
\examples{
cellosaurus <- read_cellosaurus_xml(system.file("extdata",
                                                "cellosaurus.xml",
                                                package = "rcellosaurus"))
cell_lines <- cell_lines_all(cellosaurus)

}
