% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fullpull}
\alias{fullpull}
\title{fullpull}
\usage{
fullpull(cvcl, melt.cells)
}
\arguments{
\item{cvcl}{Cellosaurus CVCL id (e.g. CVCL_1384)}

\item{melt.cells}{data.frame made from cellosaurus xml}
}
\value{
Dataframe of all synonymous (SS), similar origins (OI), MSI positive, or problematic
lines associated with the CVCL
}
\description{
Gets the synonymous cell lines and cell line that they derived from,
as well as any indication of MSI or probelmatic cell line (e.g. contamination)
}
\examples{
data(melt.cells)
cvcl <- getCVCL('Hela', melt.cells)
fullpull(cvcl, melt.cells)

fullpull("HeLa", melt.cells)
}
