% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\docType{package}
\name{rcellosaurus}
\alias{rcellosaurus}
\alias{rcellosaurus-package}
\title{rcellosaurus: A package for using the Cellosaurus dataset.}
\description{
The rcellosaurus provides functions for accessing the Cellosaurus dataset in
R. Cellosaurus is a knowledge resource on cell lines. It is online at
http://web.expasy.org/cellosaurus/
}
\details{
This R package is not associated with Cellosaurus, and the good people at
Cellosaurus are not responsible for it in any way.

The rcellosaurus package provides functions in two broad categories:
functions for finding cell-lines that match search terms and functions for
accessing information on those cell-lines. In addition,
\code{\link{read_cellosaurus_xml}} is used to read the Cellosaurus dataset.
}
\section{Finding cell-lines}{

The following functions are provided to find cell-lines:
\itemize{
  \item \code{\link{cell_line_find_first}}, to find one cell-lines
  \item \code{\link{cell_line_find_all}}, to find multiple cell-lines
}
}

\section{Accessing cell-line information}{

The following functions are provided to find cell-lines:
\itemize{
  \item \code{\link{cell_line_category}},
  \item \code{\link{cell_line_sex}},
  \item \code{\link{cell_line_names}},
}
}

