% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_find_all}
\alias{cell_line_find_all}
\title{Find all matching cell-lines}
\usage{
cell_line_find_all(cellosaurus, text)
}
\description{
Find all matching cell-lines
}
\examples{
# Finding all the cell-lines that match "sapiens".
# As of Cellosaurus v22.1 (2017-05-17) returned 69593 results.
# Takes a minute or two on a reasonably quick modern laptop.
cellosaurus <- read_cellosaurus_xml("data/cellosaurus.xml")
cell_line_find_all(cellosaurus, "sapiens")

}
