% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_find_first}
\alias{cell_line_find_first}
\title{Find cell-line}
\usage{
cell_line_find_first(cellosaurus, text)
}
\arguments{
\item{cellosaurus}{An XML document containing the Cellosaurus dataset.}

\item{text}{Some text to search for. Finding is handled using XPath so some
special characters may cause difficulties, including these: '/:[]*.
Can take a character vector of length > 1, in which case will search for
any of the terms (i.e., "or").}
}
\value{
An XML node containing a cell-line (or an \code{xml_missing} node if
  the search text is not found). If there are multiple matches, the first
  is returned.
}
\description{
Find cell-line
}
\examples{
cellosaurus <- read_cellosaurus_xml(system.file("extdata",
                                                "cellosaurus.xml",
                                                package = "rcellosaurus"))
cell_line_find_first(cellosaurus, "CVCL_3449")

# Can store the cell-line for use later.
CVCL_6873 <- cell_line_find_first(cellosaurus, "CVCL_6873")

}
