% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{cell_line_names}
\alias{cell_line_names}
\title{Find names from cell-lines}
\usage{
cell_line_names(cell_line, type = NULL)
}
\arguments{
\item{cell_line}{A node or nodeset referring to a cell-line or multiple cell-lines.}

\item{type}{A character vector specifying which types to output. If \code{NULL},
the function returns all values (i.e., unfiltered).}
}
\value{
The name(s) associated with the cell-line.
}
\description{
Expected values for \code{type} are \code{c("identifier", "synonym")}.
}
\examples{
cellosaurus <- read_cellosaurus_xml("data/cellosaurus.xml")
a_line <- cell_line_find_first(cellosaurus, "CVCL_E548")

# If type is not set, returns all names
cell_line_names(a_line)

# The expected values of type are "identifier" or "synonym":
cell_line_names(a_line, type = "identifier")
cell_line_names(a_line, type = "synonym")

# Can provide multiple type values as a character vector.
# This should be the same as the unfiltered version:
cell_line_names(a_line, type = c("identifier", "synonym"))

}
