% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_line-details.R
\name{cell_line_sex}
\alias{cell_line_sex}
\title{Sex of a cell-line}
\usage{
cell_line_sex(cell_line)
}
\arguments{
\item{cell_line}{A node or nodeset referring to a cell-line.}
}
\value{
The sex attribute associated with the cell-line. \code{NA} if
  no sex attribute is present.
}
\description{
Sex of a cell-line
}
\examples{
# First read the dataset in.
cellosaurus <- read_cellosaurus_xml(system.file("extdata",
                                                "cellosaurus.xml",
                                                package = "rcellosaurus"))

# Can call on a nodeset that's already been found. E.g.:
a_line <- cell_line_find_first(cellosaurus, "CVCL_E548")
cell_line_sex(a_line)

# Works with nodesets containing multiple cell-lines:
mice_lines <- cell_line_find_all(cellosaurus, "Mus musculus")
cell_line_sex(mice_lines)

# Can nest the sex function around the the finder function:
cell_line_sex(cell_line_find_first(cellosaurus, "CVCL_3449"))

}
