% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus.R
\name{read_cellosaurus_xml}
\alias{read_cellosaurus_xml}
\title{Load a Cellosaurus xml dataset}
\usage{
read_cellosaurus_xml(data)
}
\arguments{
\item{data}{A Cellosaurus XML data file}
}
\value{
An XML document object.
}
\description{
Cellosaurus publishes its dataset as an XML file, available for download
from ftp://ftp.expasy.org/databases/cellosaurus
}
\examples{
cellosaurus <- read_cellosaurus_xml(system.file("extdata",
                                                "cellosaurus.xml",
                                                package = "rcellosaurus"))

}
